requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
boto3==1.35.49
botocore==1.35.49
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langsmith==0.1.137
marshmallow==3.23.0
momento==1.24.0
momento-wire-types==0.106.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.52.2
orjson==3.10.10
packaging==24.1
propcache==0.2.0
protobuf==4.25.5
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.3
six==1.16.0
slack_bolt==1.21.2
slack_sdk==3.33.2
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
